refactor(sandbox): extract fake client
